Add hierarchical options.
There are now 2 levels of options in the app.

The first option, serivce_type, determines if Moodica acts like a search engine or a music recommender.
The values for this first option can be "Search" or "Recommendation".
If the first level option is "Search", the second level options are:

1. 'Artists', 2. 'Albums', 3. 'Songs', and 4. 'Lyrics'

If the first level option is "Recommendation", the seond level options are:

1. 'Song' and 2. 'Playlist'.